backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
python-jose[cryptography]
passlib
bcrypt==4.0.1
pillow
python-socketio
email-validator
transformers
dateparser